legacy/backend_fastapi_archive/fastapi_app/auth.py
"""Authentication utilities: password hashing, JWT creation & decoding."""

from datetime import datetime, timedelta
from typing import Optional
import os
from jose import jwt, JWTError
from passlib.context import CryptContext
import re

ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Pull secret key from env; fall back to a clearly insecure default with warning
SECRET_KEY = os.getenv("FASTAPI_SECRET_KEY", "INSECURE_DEMO_KEY_CHANGE_ME")

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)

def create_access_token(subject: str, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = {"sub": subject}
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def decode_access_token(token: str) -> Optional[str]:
    """Return subject (username) if valid token else None."""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload.get("sub")
    except JWTError:
        return None


def validate_password_strength(password: str) -> Optional[str]:
    """Return None if password strong enough else error message.
    Current policy (lightweight): length >=8, at least one letter & one digit.
    """
    if len(password) < 8:
        return "Password must be at least 8 characters"
    if not re.search(r"[A-Za-z]", password):
        return "Password must contain a letter"
    if not re.search(r"\d", password):
        return "Password must contain a digit"
    return None

