legacy/backend_fastapi_archive/requirements.txt
# Consolidated requirements for mixed Django (legacy) + FastAPI + ML stack.
# NOTE: Avoid conflicting duplicate pins. Django kept at 4.2 LTS for stability.

# Django (legacy components retained temporarily)
Django==4.2.24
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.8.0
drf-spectacular==0.28.0
python-dotenv==1.1.1

# Database
psycopg2-binary>=2.9.0

# ML and Data Science
numpy==2.3.3
pandas==2.3.2
scikit-learn==1.7.2
onnxruntime==1.22.1
shap==0.48.0
joblib==1.5.2

# ONNX conversion
skl2onnx==1.19.1
onnx==1.19.0
onnxmltools==1.14.0

# File processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Development and Testing
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0
factory-boy==3.3.3

# Production
gunicorn==23.0.0
whitenoise==6.10.0

# Monitoring and Logging
sentry-sdk==2.38.0

# Utils
requests==2.32.5
celery==5.5.3
redis==6.4.0

biopython==1.83
coverage==7.4.1

# FastAPI stack and MLFlow
fastapi==0.115.12
uvicorn[standard]==0.30.6
sqlalchemy==2.0.41
pydantic==2.11.7
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
mlflow==2.15.1
torch==2.4.1